Title: 최소 에너지 비용, Time: 152ms, Memory: 13MB, Status: Passed - CodeTree

+++ 240408/최소 에너지 비용/minimum-energy-cost.java
@@ -31,9 +31,9 @@ public class Main {
 		}
 //		System.out.println(Arrays.toString(minCost));
 //		System.out.println(Arrays.toString(moveCost));
-		int answer = 0;
+		long answer = 0;
 		for (int i = 0; i < N - 1; i++) {
-			answer += minCost[i] * moveCost[i];
+			answer += (long) minCost[i] * moveCost[i];
 		}
 
 		System.out.println(answer);
